requirements.txt
fastapi
uvicorn
rembg
pillow
python-multipart
onnxruntime